requirements.txt
sentence-transformers
numpy
streamlit
sentence-transformers
numpy
streamlit
groq
python-dotenv